hackbackend/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.11
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.2
django-environ==0.8.1
djangorestframework==3.13.1
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pyparsing==3.0.7
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
